+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 [Unreleased]: https://github.com/mickvangelderen/gl-typed-rust/compare/v0.4.3...HEAD
 
+### Added
+ - glClearDepth
 ## [0.4.3] - 2019-04-10
 [0.4.3]: https://github.com/mickvangelderen/gl-typed-rust/compare/v0.4.2...0.4.3