+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - glClearDepth
  - Renderbuffer functionality
+ 
+### Changed
+ - Removed unused type parameter from `delete_buffers`.
 
 ## [0.4.3] - 2019-04-10
 [0.4.3]: https://github.com/mickvangelderen/gl-typed-rust/compare/v0.4.2...0.4.3