+++ CHANGELOG.md
@@ -6,11 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-[Unreleased]: https://github.com/mickvangelderen/gl-typed-rust/compare/v0.4.3...HEAD
+[Unreleased]: https://github.com/mickvangelderen/gl-typed-rust/compare/v0.4.4...HEAD
+
+## [0.4.4] - 2019-04-25
+[0.4.4]: https://github.com/mickvangelderen/gl-typed-rust/compare/v0.4.3...0.4.4
 
 ### Added
  - glClearDepth
  - Renderbuffer functionality
+ - glColorMask
+ - glDepthMask
+ - glDepthFunc
+ - glDepthRange
+ - glStencilMask
  
 ### Changed
  - Removed unused type parameter from `delete_buffers`.